dashboard/requirements.txt
streamlit==1.39.0
pandas==2.2.3
pymongo==4.10.1
cassandra-driver==3.30.0
gevent==25.9.1
folium==0.17.0
streamlit-folium==0.23.1
plotly==5.24.1
pyhive==0.7.0
thrift==0.21.0
thrift-sasl==0.4.3
pure-sasl==0.6.2
praw==7.8.1
requests==2.32.3
python-dotenv==1.0.1
